ref(slack): Isolate Seer agent actions
