+++ src/sentry/integrations/slack/webhooks/actions/common.py
@@ -1,0 +1,11 @@
+DEFAULT_ERROR_MESSAGE = "Sentry can't perform that action right now on your behalf!"
+LINK_IDENTITY_MESSAGE = (
+    "Looks like you haven't linked your Sentry account with your Slack identity yet! "
+    "<{associate_url}|Link your identity now> to perform actions in Sentry through Slack. "
+)
+NO_IDENTITY_MESSAGE = "Identity not linked for user."
+UNLINK_IDENTITY_MESSAGE = (
+    "Looks like this Slack identity is linked to the Sentry user *{user_email}* "
+    "who is not a member of organization *{org_name}* used with this Slack integration. "
+    "<{associate_url}|Unlink your identity now>. "
+)
